printf fonctionne imo
