+++ Printf/Makefile
@@ -6,7 +6,7 @@
 #    By: mathis <mathis@student.42.fr>              +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2021/11/18 19:28:02 by mathis            #+#    #+#              #
-#    Updated: 2022/01/25 12:03:24 by mathis           ###   ########.fr        #
+#    Updated: 2022/03/04 12:31:43 by mathis           ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -14,54 +14,24 @@
 CC = gcc
 NAME = libftprintf.a
 
-SRC = ft_atoi.c \
-	  ft_bzero.c \
-	  ft_calloc.c \
-	  ft_isalnum.c \
-	  ft_isalpha.c \
-	  ft_isascii.c \
-	  ft_isdigit.c \
-	  ft_isprint.c \
-	  ft_itoa.c \
-	  ft_memchr.c \
-	  ft_memcmp.c \
-	  ft_memcpy.c \
-	  ft_memmove.c \
-	  ft_memset.c \
-	  ft_putchar_fd.c \
-	  ft_putendl_fd.c \
-	  ft_putnbr_fd.c \
-	  ft_putstr_fd.c \
-	  ft_split.c \
-	  ft_strchr.c \
-	  ft_strdup.c \
-	  ft_striteri.c \
-	  ft_strjoin.c \
-	  ft_strlcat.c \
-	  ft_strlcpy.c \
-	  ft_strlen.c \
-	  ft_strmapi.c \
-	  ft_strncmp.c \
-	  ft_strnstr.c \
-	  ft_strrchr.c \
-	  ft_strtrim.c \
-	  ft_substr.c \
-	  ft_tolower.c \
-	  ft_toupper.c
+SRC = ft_is_arg.c \
+	  ft_is_c.c \
+	  ft_is_s.c \
+	  ft_is_p.c \
+	  ft_is_d.c \
+	  ft_is_i.c \
+	  ft_is_u.c \
+	  ft_is_x.c \
+	  ft_is_majx.c \
+	  ft_is_pcent.c \
+	  ft_putnbr_base.c \
+	  ft_putnbr_base_ui.c \
+	  ft_putnbr_basev.c \
+	  ft_strlen_unsigned.c \
+	  ft_printf.c \
 
-SRC_BONUS = $(SRC) \
-			ft_lstnew.c \
-			ft_lstsize.c \
-			ft_lstadd_front.c \
-			ft_lstlast.c \
-			ft_lstadd_back.c \
-			ft_lstdelone.c \
-			ft_lstclear.c \
-			ft_lstiter.c \
-			ft_lstmap.c
 
 OBJ = $(SRC:.c=.o)
-OBJ_BONUS = $(SRC_BONUS:.c=.o)
 
 CFLAGS = -Wall -Wextra -Werror
 RM = rm -f
@@ -77,9 +47,6 @@ all : $(NAME)
 $(NAME) : $(OBJ)
 	ar rcs $@ $^
 
-bonus : $(OBJ_BONUS)
-	ar rcs $(NAME) $^
-
 clean :
 	$(RM) *.o
 
@@ -88,4 +55,4 @@ fclean : clean
 
 re : fclean all
 
-.PHONY : all clean fclean re bonus
+.PHONY : all clean fclean re
