+++ Printf/ft_is_i.c
@@ -1,27 +1,31 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_striteri.c                                      :+:      :+:    :+:   */
+/*   ft_is_i.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mravera <@student.42lausanne.ch>           +#+  +:+       +#+        */
+/*   By: mathis <mathis@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2021/12/09 18:21:10 by mravera           #+#    #+#             */
-/*   Updated: 2021/12/09 19:17:16 by mravera          ###   ########.fr       */
+/*   Created: 2022/03/04 11:36:51 by mathis            #+#    #+#             */
+/*   Updated: 2022/03/04 12:11:34 by mathis           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libftprintf.h"
 
-void	ft_striteri(char *s, void (*f)(unsigned int, char*))
+int	is_i(va_list lst)
 {
+	int		x;
+	char	*res;
 	size_t	i;
 
-	if (s == NULL)
-		return ;
 	i = 0;
-	while (s[i])
+	res = (char *)malloc(34 * sizeof(char));
+	x = va_arg(lst, const int);
+	ft_putnbr_base(res, x, "0123456789");
+	while (res[i])
 	{
-		f(i, &(s[i]));
+		write(1, &res[i], 1);
 		i++;
 	}
+	return (i);
 }
