+++ Printf/ft_is_p.c
@@ -1,33 +1,33 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_strdup.c                                        :+:      :+:    :+:   */
+/*   ft_is_p.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mathis <marvin@42.fr>                      +#+  +:+       +#+        */
+/*   By: mathis <mathis@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2021/12/02 14:11:03 by mathis            #+#    #+#             */
-/*   Updated: 2022/01/12 14:08:33 by mathis           ###   ########.fr       */
+/*   Created: 2022/03/04 11:37:04 by mathis            #+#    #+#             */
+/*   Updated: 2022/03/04 12:09:57 by mathis           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libftprintf.h"
-#include <stdlib.h>
 
-char	*ft_strdup(const char *s)
+int	is_p(va_list lst)
 {
-	char	*res;
+	void	*x;
 	size_t	i;
-	size_t	len;
+	char	*res;
 
 	i = 0;
-	len = ft_strlen(s);
-	res = malloc(len + 1);
-	if (res == 0)
-		return (0);
-	while (i <= ft_strlen(s))
+	res = (char *)malloc(34 * sizeof(char));
+	x = va_arg(lst, void *);
+	ft_putnbr_basev(res, (unsigned long long) x, "0123456789abcdef");
+	write(1, "0", 1);
+	write(1, "x", 1);
+	while (res[i])
 	{
-		res[i] = s[i];
+		write(1, &res[i], 1);
 		i++;
 	}
-	return (res);
+	return (i + 2);
 }
