--- Printf/ft_isalnum.c
@@ -1,34 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   ft_isalnum.c                                       :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: mravera <@student.42lausanne.ch>           +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2021/10/27 12:20:58 by mravera           #+#    #+#             */
-/*   Updated: 2021/11/24 17:57:32 by mravera          ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "libftprintf.h"
-
-int	ft_isalnum(int c)
-{
-	if ((c >= '0' && c <= '9')
-		|| (c >= 'A' && c <= 'Z')
-		|| (c >= 'a' && c <= 'z' ))
-		return (1);
-	else
-		return (0);
-}
-
-/*
-#include <stdio.h>
-
-int	main(int argc, char **argv)
-{
-	(void) argc;
-	printf("%i ==> %i", argv[1][0], ft_isalnum(argv[1][0]));
-	return (0);
-}
-*/
