--- Printf/ft_lstadd_front.c
@@ -1,19 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   ft_lstadd_front.c                                  :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: mravera <@student.42lausanne.ch>           +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2021/12/13 20:08:15 by mravera           #+#    #+#             */
-/*   Updated: 2021/12/13 20:20:46 by mravera          ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "libftprintf.h"
-
-void	ft_lstadd_front(t_list **alst, t_list *new)
-{
-	new->next = *alst;
-	*alst = new;
-}
